test: Split unittests for utils module
