+++ test/test_utils_geometry.py
@@ -1,0 +1,13 @@
+from doctr.utils import geometry
+
+
+def test_bbox_to_polygon():
+    assert geometry.bbox_to_polygon(((0, 0), (1, 1))) == ((0, 0), (1, 0), (0, 1), (1, 1))
+
+
+def test_polygon_to_bbox():
+    assert geometry.polygon_to_bbox(((0, 0), (1, 0), (0, 1), (1, 1))) == ((0, 0), (1, 1))
+
+
+def test_resolve_enclosing_bbox():
+    assert geometry.resolve_enclosing_bbox([((0, 0.5), (1, 0)), ((0.5, 0), (1, 0.25))]) == ((0, 0), (1, 0.5))
